Remove NotificationService
