+++ lib/main.dart
@@ -8,15 +8,11 @@ import 'core/theme.dart';
 import 'providers/news_provider.dart';
 import 'providers/settings_provider.dart';
 import 'providers/auth_provider.dart';
-import 'services/firebase_service.dart';
-import 'services/notification_service.dart';
 
 
 void main() async {
   WidgetsFlutterBinding.ensureInitialized();
 
-  await FirebaseService().initialize();
-  await NotificationService().initialize();
 
   // Initialize Mobile Ads
   // MobileAds.instance.initialize();
